requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.25

# Sentiment Analysis
transformers==4.37.0
torch==2.2.0
sentencepiece==0.1.99

# Scheduling
apscheduler==3.10.4

# Dashboard
streamlit==1.31.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2

# Testing (optional)
pytest==8.0.0
pytest-cov==4.1.0
